+++ CHANGELOG.md
@@ -14,3 +14,9 @@ All notable changes to the "work-progress" extension will be documented in this 
 - Emails sent after each session (have to enable, sorry)
 - Time measured without having to enter commands
 - API is now manageble from settings
+
+## 1.1.0 (still in dev)
+
+- Checks idle time
+- Sends email if user not focused for 1 hour and stops the timer
+- Updates user settings on activation